requirements.txt
# Core web
pydantic>=2.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# OpenEnv spec
openenv-core>=0.1.0

# Numerics
numpy>=1.26.0

# LLM client (inference.py)
openai>=1.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
